+++ capa/ghidra/capa_ghidra.py
@@ -160,12 +160,8 @@ def main():
 
 
 if __name__ == "__main__":
-    if sys.version_info < (3, 8):
+    if sys.version_info < (3, 10):
         from capa.exceptions import UnsupportedRuntimeError
 
-        raise UnsupportedRuntimeError("This version of capa can only be used with Python 3.8+")
-    elif sys.version_info < (3, 10):
-        from warnings import warn
-
-        warn("This is the last capa version supporting Python 3.8 and 3.9.", DeprecationWarning, stacklevel=2)
+        raise UnsupportedRuntimeError("This version of capa can only be used with Python 3.10+")
     sys.exit(main())